file : play파일들 이동
